# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GridScore AI API",
    "summary": "Credit infrastructure for Africa's energy lenders.",
    "description": "Cooperative PAYG repayment data platform. **All data in non-production environments is synthetic.**",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/operators": {
      "get": {
        "tags": [
          "operators"
        ],
        "summary": "List operators",
        "operationId": "list_operators_v1_operators_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OperatorOut"
                  },
                  "type": "array",
                  "title": "Response List Operators V1 Operators Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "operators"
        ],
        "summary": "Onboard an operator",
        "operationId": "create_operator_v1_operators_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperatorCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/operators/{operator_id}/api-keys": {
      "post": {
        "tags": [
          "operators"
        ],
        "summary": "Issue a machine API key for an operator",
        "operationId": "issue_api_key_v1_operators__operator_id__api_keys_post",
        "parameters": [
          {
            "name": "operator_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Operator Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/operators/users": {
      "post": {
        "tags": [
          "operators"
        ],
        "summary": "Create a user account",
        "operationId": "create_user_v1_operators_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/score": {
      "post": {
        "tags": [
          "scoring"
        ],
        "summary": "Score a customer (solo or pooled view)",
        "operationId": "score_v1_score_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/score/cooperative": {
      "post": {
        "tags": [
          "scoring"
        ],
        "summary": "Cooperative lift: solo vs pooled, with the decision flip",
        "operationId": "score_cooperative_v1_score_cooperative_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CooperativeScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CooperativeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiKeyCreated": {
        "properties": {
          "prefix": {
            "type": "string",
            "title": "Prefix"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "description": "Full key, shown once. Store it securely."
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": "This is the only time the full key is shown."
          }
        },
        "type": "object",
        "required": [
          "prefix",
          "api_key"
        ],
        "title": "ApiKeyCreated"
      },
...
      "CooperativeOut": {
        "properties": {
          "customer_id": {
            "type": "string",
            "format": "uuid",
            "title": "Customer Id"
          },
          "solo": {
            "$ref": "#/components/schemas/ScoreOut"
          },
          "pooled": {
            "$ref": "#/components/schemas/ScoreOut"
          },
          "pd_delta": {
            "type": "number",
            "title": "Pd Delta"
          },
          "confidence_delta": {
            "type": "number",
            "title": "Confidence Delta"
          },
          "score_delta": {
            "type": "integer",
            "title": "Score Delta"
          },
          "decision_flips": {
            "type": "boolean",
            "title": "Decision Flips"
          },
          "lift_metric": {
            "type": "number",
            "title": "Lift Metric"
          }
        },
        "type": "object",
        "required": [
          "customer_id",
          "solo",
          "pooled",
          "pd_delta",
          "confidence_delta",
          "score_delta",
          "decision_flips",
          "lift_metric"
        ],
        "title": "CooperativeOut"
      },
      "CooperativeScoreRequest": {
        "properties": {
          "customer_id": {
            "type": "string",
            "format": "uuid",
            "title": "Customer Id"
          }
        },
        "type": "object",
        "required": [
          "customer_id"
        ],
        "title": "CooperativeScoreRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OperatorCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Name"
          },
          "country": {
            "type": "string",
            "maxLength": 2,
            "minLength": 2,
            "title": "Country",
            "description": "ISO 3166-1 alpha-2"
          }
        },
        "type": "object",
        "required": [
          "name",
          "country"
        ],
        "title": "OperatorCreate"
      },
      "OperatorOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "country": {
            "type": "string",
            "title": "Country"
          },
          "status": {
            "$ref": "#/components/schemas/OperatorStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "country",
          "status",
          "created_at"
        ],
        "title": "OperatorOut"
      },
      "OperatorStatus": {
        "type": "string",
        "enum": [
          "pending",
          "active",
          "suspended"
        ],
        "title": "OperatorStatus",
        "description": "Lifecycle of a tenant (PAYG operator) in the cooperative."
      },
...
      "RiskTier": {
        "type": "string",
        "enum": [
          "A",
          "B",
          "C",
          "D",
          "E"
        ],
        "title": "RiskTier",
        "description": "Risk grade with documented PD cutoffs (defined in the scoring service)."
      },
...
      "ScoreOut": {
        "properties": {
          "customer_id": {
            "type": "string",
            "format": "uuid",
            "title": "Customer Id"
          },
          "view": {
            "$ref": "#/components/schemas/ScoreView"
          },
          "default_probability": {
            "type": "number",
            "title": "Default Probability"
          },
          "energy_credit_score": {
            "type": "integer",
            "title": "Energy Credit Score"
          },
          "risk_tier": {
            "$ref": "#/components/schemas/RiskTier"
          },
          "approved": {
            "type": "boolean",
            "title": "Approved"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "top_factors": {
            "items": {
              "$ref": "#/components/schemas/TopFactorOut"
            },
            "type": "array",
            "title": "Top Factors"
          }
        },
        "type": "object",
        "required": [
          "customer_id",
          "view",
          "default_probability",
          "energy_credit_score",
          "risk_tier",
          "approved",
          "model_version",
          "top_factors"
        ],
        "title": "ScoreOut"
      },
      "ScoreRequest": {
        "properties": {
          "customer_id": {
            "type": "string",
            "format": "uuid",
            "title": "Customer Id"
          },
          "view": {
            "$ref": "#/components/schemas/ScoreView",
            "default": "pooled"
          }
        },
        "type": "object",
        "required": [
          "customer_id"
        ],
        "title": "ScoreRequest"
      },
      "ScoreView": {
        "type": "string",
        "enum": [
          "solo",
          "pooled"
        ],
        "title": "ScoreView",
        "description": "The two cooperative views \u2014 the core differentiator.\n\n``SOLO``   \u2014 features from the home operator's partial history only.\n``POOLED`` \u2014 features from the full cooperative history."
      },
...
      "TopFactorOut": {
        "properties": {
          "feature": {
            "type": "string",
            "title": "Feature"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "contribution": {
            "type": "number",
            "title": "Contribution"
          },
          "direction": {
            "type": "string",
            "title": "Direction"
          }
        },
        "type": "object",
        "required": [
          "feature",
          "label",
          "value",
          "contribution",
          "direction"
        ],
        "title": "TopFactorOut"
      },
      "UserCreate": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 320,
            "minLength": 3,
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "operator_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator Id"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "role"
        ],
        "title": "UserCreate"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "operator_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "role",
          "operator_id"
        ],
        "title": "UserOut"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "platform_admin",
          "operator_admin",
          "operator_analyst",
          "lender_viewer"
        ],
        "title": "UserRole",
        "description": "RBAC roles. Drives authorisation in Stage 5."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}